wip: split up asm module
